Setup Post Routes
Implemented same functionalities for Posts as done previously for Users.  createPost is a Create Function that creates the posts. getFeedPosts() and getUserPosts() are both Read functions . likePost() is an update function. All functions that return data are formatted and passed so that the frontend can use them and the frontend can be updated to show new and updated data such as number of likes.